Forge-Insight/requirements.txt
beautifulsoup4==4.12.3
bech32==1.2.0
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
codebase-to-text==1.0.6
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
DateTime==5.5
et-xmlfile==1.1.0
filelock==3.13.4
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
httplib2==0.22.0
idna==3.7
iniconfig==2.0.0
kiwisolver==1.4.5
lxml==5.2.1
matplotlib==3.9.2
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.1
pandas==2.2.0
pillow==10.4.0
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.27.3
pyaes==1.6.1
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyparsing==3.1.2
PySocks==1.7.1
pyTelegramBotAPI==4.17.0
pytest==8.3.2
python-dateutil==2.8.2
python-docx==1.1.2
pytz==2024.1
requests==2.32.2
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
seaborn==0.13.2
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
snscrape==0.7.0.20230622
soupsieve==2.5
telebot==0.0.5
Telethon==1.35.0
tqdm==4.66.3
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.2
wheel==0.44.0
zope.interface==6.3
